Factorize generic code from ExternalDictLibBridgeHelper into base class
- When catboost will be added into library-bridge later on, a lot of
  code will need to be duplicated.

- Avoid that by factorizing stuff which is not specific to external
  dictionaries for reuse into a common base class.

- This is a refactoring without semantic change.
